rename abstract*
